:recycle: Rename main writer to -> RagnarokRewriter
